break up loba.rb
